refactor(sprint9): main.rs (1556 LOC) split — 5 helper modules
- event_handlers.rs (131) : persist_event, emit_event, api_events, api_events_stream
- oauth_utils.rs (63) : normalize_segment, build_misfits_local, normalize_oauth_provider
- mailing_list_handlers.rs (118) : create_mailing_list, send_to_mailing_list
- send_dispatch.rs (80) : send_email_handler
- web_utils.rs (81) : req_ip_str, get_accept_language, welcome_email_html

main.rs : 1556 → 1119 LOC (main() + routes seuls)
