+++ src/bin/email_api_dir/main.rs
@@ -209,459 +209,22 @@ struct AuthResponse {
 
 
 
-async fn persist_event(mongo: &mongodb::Client, event: &MailEvent) {
-    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("mail_events");
-    if let Ok(doc) = bson::to_document(event) {
-        if let Err(e) = coll.insert_one(doc).await {
-            eprintln!("persist_event error: {}", e);
-        }
-    }
-}
 
-async fn emit_event(bus: &EventBus, mongo: &mongodb::Client, event: MailEvent) {
-    persist_event(mongo, &event).await;
-    let _ = bus.send(event);
-}
+mod event_handlers;
 
-async fn api_events(
-    req: actix_web::HttpRequest,
-    mongo: web::Data<Arc<mongodb::Client>>,
-) -> impl Responder {
-    let user_id = resolve_user_id(&req);
-    let db = env::var("MONGODB_DATABASE").unwrap_or_else(|_| "mailserver".to_string());
-    let coll = mongo
-        .database(&db)
-        .collection::<bson::Document>("mail_events");
-    let filter = doc! { "user_id": &user_id };
-    match coll
-        .find(filter)
-        .sort(doc! { "timestamp": -1 })
-        .limit(200)
-        .await
-    {
-        Ok(cursor) => match cursor.try_collect::<Vec<_>>().await {
-            Ok(docs) => {
-                let events: Vec<serde_json::Value> = docs
-                    .into_iter()
-                    .filter_map(|d| bson::from_document::<MailEvent>(d).ok())
-                    .filter_map(|e| serde_json::to_value(&e).ok())
-                    .collect();
-                HttpResponse::Ok().json(serde_json::json!({ "events": events }))
-            }
-            Err(e) => HttpResponse::InternalServerError()
-                .json(serde_json::json!({ "error": e.to_string() })),
-        },
-        Err(e) => {
-            HttpResponse::InternalServerError().json(serde_json::json!({ "error": e.to_string() }))
-        }
-    }
-}
+mod oauth_utils;
 
-async fn api_events_stream(bus: web::Data<EventBus>, req: actix_web::HttpRequest) -> HttpResponse {
-    let user_id = resolve_user_id(&req);
-    let rx = bus.subscribe();
-    let event_stream = stream::unfold((rx, user_id), |(mut rx, uid)| async move {
-        loop {
-            match rx.recv().await {
-                Ok(event) if event.user_id == uid => {
-                    let data = serde_json::to_string(&event).unwrap_or_default();
-                    let chunk = format!("data: {}\n\n", data);
-                    return Some((
-                        Ok::<web::Bytes, actix_web::Error>(web::Bytes::from(chunk)),
-                        (rx, uid),
-                    ));
-                }
-                Ok(_) => continue,
-                Err(broadcast::error::RecvError::Lagged(_)) => continue,
-                Err(broadcast::error::RecvError::Closed) => return None,
-            }
-        }
-    });
-    HttpResponse::Ok()
-        .content_type("text/event-stream")
-        .insert_header(("Cache-Control", "no-cache"))
-        .insert_header(("X-Accel-Buffering", "no"))
-        .streaming(event_stream)
-}
+mod mailing_list_handlers;
 
-// ===========================================================================
-// SMTP Monitoring endpoints
-// ===========================================================================
+mod send_dispatch;
 
-use simple_smtp_server::monitoring;
-use simple_smtp_server::monitoring::alerts::AlertConfig;
-use simple_smtp_server::monitoring::storage;
-use simple_smtp_server::security;
+mod web_utils;
+pub use event_handlers::*;
+pub use oauth_utils::*;
+pub use mailing_list_handlers::*;
+pub use send_dispatch::*;
+pub use web_utils::*;
 
-// parse_window, since_str, env_bool, monitoring query types, dns_txt_lookup
-// api_monitoring_*, SecurityAlertsQuery, api_security_*
-// → moved to monitoring_handlers module
-
-
-
-// ↑ monitoring/security handlers moved to monitoring_handlers module ↑
-
-
-
-#[derive(Deserialize)]
-struct DeliverabilityDiagnosticsQuery {
-    #[serde(default = "default_window")]
-    window: String,
-    domain: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct DeliverabilityProcedureUpdateRequest {
-    checklist: Option<Vec<DeliverabilityChecklistUpdate>>,
-    reminder: Option<DeliverabilityReminderUpdate>,
-}
-
-#[derive(Deserialize)]
-struct DeliverabilityChecklistUpdate {
-    id: String,
-    checked: bool,
-    note: Option<String>,
-}
-
-#[derive(Deserialize)]
-struct DeliverabilityReminderUpdate {
-    enabled: bool,
-    cadence_hours: u32,
-}
-
-/// GET /api/monitoring/summary?window=15m
-// monitoring + security handlers → monitoring_handlers module
-
-
-fn normalize_segment(s: &str) -> String {
-    s.chars()
-        .map(|c| match c {
-            'à' | 'â' | 'ä' => 'a',
-            'é' | 'è' | 'ê' | 'ë' => 'e',
-            'î' | 'ï' => 'i',
-            'ô' | 'ö' => 'o',
-            'ù' | 'û' | 'ü' => 'u',
-            'ç' => 'c',
-            'ñ' => 'n',
-            'æ' => 'a',
-            'œ' => 'o',
-            _ => c,
-        })
-        .filter(|c| c.is_ascii_alphanumeric())
-        .map(|c| c.to_ascii_lowercase())
-        .collect()
-}
-
-/// Retourne `prenom.nom` normalisé, ou None si prénom ou nom est absent.
-fn build_misfits_local(first_name: &str, last_name: &str) -> Option<String> {
-    let first = normalize_segment(first_name.trim());
-    let last = normalize_segment(last_name.trim());
-    if first.is_empty() || last.is_empty() {
-        return None;
-    }
-    Some(format!("{}.{}", first, last))
-}
-
-fn normalize_oauth_provider(provider: &str) -> Option<String> {
-    let p = provider.trim().to_ascii_lowercase();
-    match p.as_str() {
-        "github" => Some(p),
-        _ => None,
-    }
-}
-
-#[derive(Deserialize, Serialize, Debug, PartialEq)]
-pub struct EmailRequest {
-    from: String,
-    to: String,
-    subject: String,
-    body: String,
-}
-
-#[derive(Deserialize)]
-struct MailingListRequest {
-    label: String,
-    emails: Vec<String>,
-}
-
-#[derive(Deserialize)]
-struct MailingListEmailRequest {
-    from: String,
-    subject: String,
-    body: String,
-    mailing_list: String,
-}
-
-async fn create_mailing_list(mailing_list: web::Json<MailingListRequest>) -> impl Responder {
-    let mailing_list_dir = Path::new("mailing-lists");
-    if !mailing_list_dir.exists() {
-        if let Err(e) = create_dir_all(mailing_list_dir) {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to create mailing list directory: {}", e)
-            }));
-        }
-    }
-
-    let file_path = mailing_list_dir.join(format!("{}.csv", mailing_list.label));
-    let mut file = match File::create(&file_path) {
-        Ok(file) => file,
-        Err(e) => {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to create file: {}", e)
-            }))
-        }
-    };
-
-    for email in &mailing_list.emails {
-        if let Err(e) = writeln!(file, "{}", email) {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to write to file: {}", e)
-            }));
-        }
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "status": "success",
-        "message": format!("Mailing list '{}' created successfully", mailing_list.label)
-    }))
-}
-
-async fn send_to_mailing_list(email_req: web::Json<MailingListEmailRequest>) -> impl Responder {
-    let mailing_list_path =
-        Path::new("mailing-lists").join(format!("{}.csv", email_req.mailing_list));
-
-    if !mailing_list_path.exists() {
-        return HttpResponse::NotFound().json(serde_json::json!({
-            "status": "error",
-            "message": format!("Mailing list '{}' not found", email_req.mailing_list)
-        }));
-    }
-
-    let file = match File::open(&mailing_list_path) {
-        Ok(file) => file,
-        Err(e) => {
-            return HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": format!("Failed to open mailing list file: {}", e)
-            }))
-        }
-    };
-
-    let reader = BufReader::new(file);
-    let mut success_count = 0;
-    let mut failure_count = 0;
-
-    for line in reader.lines() {
-        let to_email = match line {
-            Ok(email) => email.trim().to_string(),
-            Err(_e) => {
-                failure_count += 1;
-                continue;
-            }
-        };
-
-        let from = email_req.from.clone();
-        let to = to_email;
-        let subject = email_req.subject.clone();
-        let body = email_req.body.clone();
-
-        let _email_content = format!(
-            "From: {}\r\nTo: {}\r\nSubject: {}\r\n\r\n{}",
-            from, to, subject, body
-        );
-        let client = reqwest::Client::new();
-        let dkim_service_url = env::var("DKIM_SERVICE_URL").expect("DKIM_SERVICE_URL not set");
-
-        let _dkim_response = match client
-            .post(&dkim_service_url)
-            .json(&serde_json::json!({
-                "from": from,
-                "to": to,
-                "subject": subject,
-                "text": body
-            }))
-            .send()
-            .await
-        {
-            Ok(resp) => {
-                if resp.status().is_success() {
-                    success_count += 1;
-                } else {
-                    failure_count += 1;
-                }
-            }
-            Err(_) => {
-                failure_count += 1;
-            }
-        };
-    }
-
-    HttpResponse::Ok().json(serde_json::json!({
-        "status": "success",
-        "message": format!("Emails sent to mailing list '{}'. Successful: {}, Failed: {}",
-                           email_req.mailing_list, success_count, failure_count)
-    }))
-}
-
-async fn send_email_handler(
-    email_req: web::Json<EmailRequest>,
-    dkim_service: web::Data<Box<dyn DkimService>>,
-) -> impl Responder {
-    println!("Received email request");
-
-    match dkim_service.sign_email(&email_req).await {
-        Ok(dkim_result) => {
-            println!("DKIM service returned success");
-            let message_id = dkim_result["messageId"].as_str().unwrap_or("");
-            match dkim_result["status"].as_str() {
-                Some("success") => {
-                    let sig = dkim_result["dkimSignature"]
-                        .as_str()
-                        .or_else(|| dkim_result["dkim_signature"].as_str())
-                        .unwrap_or("")
-                        .trim()
-                        .to_string();
-                    if sig.is_empty() {
-                        return HttpResponse::InternalServerError().json(serde_json::json!({
-                            "status": "error",
-                            "message": "DKIM service returned success without a DKIM signature",
-                        }));
-                    }
-
-                    // Construct the email with DKIM signature
-                    let email = Email {
-                        id: Uuid::new_v4().to_string(),
-                        from: email_req.from.clone(),
-                        to: email_req.to.clone(),
-                        subject: email_req.subject.clone(),
-                        body: email_req.body.clone(),
-                        headers: vec![("DKIM-Signature".to_string(), sig.clone())],
-                        flags: vec![],
-                        sequence_number: 0,
-                        uid: 0,
-                        internal_date: mongodb::bson::DateTime::from_millis(
-                            Utc::now().timestamp_millis(),
-                        ),
-                        dkim_signature: Some(sig),
-                    };
-
-                    // Send the email
-                    match send_outgoing_email(&email).await {
-                        Ok(_) => HttpResponse::Ok().json(serde_json::json!({
-                            "status": "success",
-                            "message": "Email signed and sent successfully",
-                            "messageId": message_id
-                        })),
-                        Err(e) => HttpResponse::InternalServerError().json(serde_json::json!({
-                            "status": "error",
-                            "message": format!("Failed to send email: {}", e)
-                        })),
-                    }
-                }
-                _ => {
-                    let error_message = dkim_result["message"].as_str().unwrap_or("Unknown error");
-                    HttpResponse::InternalServerError().json(serde_json::json!({
-                        "status": "error",
-                        "message": format!("Failed to sign email: {}", error_message)
-                    }))
-                }
-            }
-        }
-        Err(e) => {
-            eprintln!("Failed to call DKIM service: {}", e);
-            HttpResponse::InternalServerError().json(serde_json::json!({
-                "status": "error",
-                "message": "Failed to generate DKIM signature"
-            }))
-        }
-    }
-}
-
-// --- Auth handlers ---
-
-fn req_ip_str(req: &actix_web::HttpRequest) -> String {
-    req.connection_info()
-        .realip_remote_addr()
-        .unwrap_or("unknown")
-        .split(':')
-        .next()
-        .unwrap_or("unknown")
-        .to_string()
-}
-
-fn get_accept_language(req: &actix_web::HttpRequest) -> String {
-    req.headers()
-        .get("accept-language")
-        .and_then(|v| v.to_str().ok())
-        .unwrap_or("")
-        .to_string()
-}
-
-fn welcome_email_html(
-    locale: &str,
-    display_name: &str,
-    primary_email: &str,
-    alias_email: Option<&str>,
-) -> String {
-    let dir = if i18n::is_rtl(locale) { "rtl" } else { "ltr" };
-    let text_align = if i18n::is_rtl(locale) {
-        "right"
-    } else {
-        "left"
-    };
-    let greeting = i18n::t(locale, "email-welcome-greeting", &[("name", display_name)]);
-    let intro = i18n::t(locale, "email-welcome-intro", &[]);
-    let primary_label = i18n::t(locale, "email-welcome-primary-label", &[]);
-    let cta = i18n::t(locale, "email-welcome-cta", &[]);
-    let signature = i18n::t(locale, "email-welcome-signature", &[]);
-    let alias_row = alias_email
-        .map(|a| {
-            let lbl = i18n::t(locale, "email-welcome-alias-label", &[]);
-            let detail = i18n::t(locale, "email-welcome-alias-detail", &[]);
-            format!(
-                "<tr><td style=\"padding:4px 0\"><b>{lbl}</b> \
-                 <span class=\"addr\">{a}</span> {detail}</td></tr>"
-            )
-        })
-        .unwrap_or_default();
-    format!(
-        r#"<!DOCTYPE html>
-<html lang="{locale}" dir="{dir}">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width,initial-scale=1.0">
-  <style>
-    body{{font-family:Tahoma,Arial,sans-serif;direction:{dir};text-align:{text_align};color:#222;background:#f5f5f5;margin:0;padding:0}}
-    .wrap{{max-width:600px;margin:40px auto;background:#fff;border-radius:8px;padding:40px}}
-    .addr{{font-family:monospace;background:#f0f0f0;padding:2px 6px;border-radius:3px}}
-  </style>
-</head>
-<body><div class="wrap">
-  <h2>{greeting}</h2>
-  <p>{intro}</p>
-  <table cellpadding="0" cellspacing="0">
-    <tr><td style="padding:4px 0"><b>{primary_label}</b> <span class="addr">{primary_email}</span></td></tr>
-    {alias_row}
-  </table>
-  <p>{cta}</p>
-  <p>{signature}</p>
-</div></body>
-</html>"#,
-    )
-}
-
-// auth_login..api_password_reset_confirm → auth_handlers module
-
-// mailbox + send + drafts → mailbox_handlers module
-// admin CRUD + change-requests + deliverability → admin_ops_handlers module
-
-#[actix_web::main]
 async fn main() -> std::io::Result<()> {
     dotenv().ok();
