+++ src/bin/email_api_dir/send_dispatch.rs
@@ -1,0 +1,80 @@
+// send_dispatch.rs — split from main.rs (Sprint 9)
+#![allow(unused_imports)]
+use super::*;
+
+async fn send_email_handler(
+    email_req: web::Json<EmailRequest>,
+    dkim_service: web::Data<Box<dyn DkimService>>,
+) -> impl Responder {
+    println!("Received email request");
+
+    match dkim_service.sign_email(&email_req).await {
+        Ok(dkim_result) => {
+            println!("DKIM service returned success");
+            let message_id = dkim_result["messageId"].as_str().unwrap_or("");
+            match dkim_result["status"].as_str() {
+                Some("success") => {
+                    let sig = dkim_result["dkimSignature"]
+                        .as_str()
+                        .or_else(|| dkim_result["dkim_signature"].as_str())
+                        .unwrap_or("")
+                        .trim()
+                        .to_string();
+                    if sig.is_empty() {
+                        return HttpResponse::InternalServerError().json(serde_json::json!({
+                            "status": "error",
+                            "message": "DKIM service returned success without a DKIM signature",
+                        }));
+                    }
+
+                    // Construct the email with DKIM signature
+                    let email = Email {
+                        id: Uuid::new_v4().to_string(),
+                        from: email_req.from.clone(),
+                        to: email_req.to.clone(),
+                        subject: email_req.subject.clone(),
+                        body: email_req.body.clone(),
+                        headers: vec![("DKIM-Signature".to_string(), sig.clone())],
+                        flags: vec![],
+                        sequence_number: 0,
+                        uid: 0,
+                        internal_date: mongodb::bson::DateTime::from_millis(
+                            Utc::now().timestamp_millis(),
+                        ),
+                        dkim_signature: Some(sig),
+                    };
+
+                    // Send the email
+                    match send_outgoing_email(&email).await {
+                        Ok(_) => HttpResponse::Ok().json(serde_json::json!({
+                            "status": "success",
+                            "message": "Email signed and sent successfully",
+                            "messageId": message_id
+                        })),
+                        Err(e) => HttpResponse::InternalServerError().json(serde_json::json!({
+                            "status": "error",
+                            "message": format!("Failed to send email: {}", e)
+                        })),
+                    }
+                }
+                _ => {
+                    let error_message = dkim_result["message"].as_str().unwrap_or("Unknown error");
+                    HttpResponse::InternalServerError().json(serde_json::json!({
+                        "status": "error",
+                        "message": format!("Failed to sign email: {}", error_message)
+                    }))
+                }
+            }
+        }
+        Err(e) => {
+            eprintln!("Failed to call DKIM service: {}", e);
+            HttpResponse::InternalServerError().json(serde_json::json!({
+                "status": "error",
+                "message": "Failed to generate DKIM signature"
+            }))
+        }
+    }
+}
+
+// --- Auth handlers ---
+
